:recycle: [refactory]: Add CustomerListContext
- Add CustomerListContext & CustomerItems component
- Update CustomerList component
- Refactory Component Tree:
CustomerListContext > CustomerList > CustomerItems
